Added Discord Verify Command

+++ 1.12.2/src/main/java/com/wurmcraft/serveressentials/forge/modules/general/command/teleport/DelHomeCommand.java
@@ -1,5 +1,7 @@
 package com.wurmcraft.serveressentials.forge.modules.general.command.teleport;
 
+import static com.wurmcraft.serveressentials.forge.api.command.SECommand.COMMAND_COLOR;
+import static com.wurmcraft.serveressentials.forge.api.command.SECommand.COMMAND_INFO_COLOR;
 import static com.wurmcraft.serveressentials.forge.api.command.SECommand.ERROR_COLOR;
 
 import com.wurmcraft.serveressentials.core.api.command.Command;
@@ -12,6 +14,7 @@ import com.wurmcraft.serveressentials.core.registry.SERegistry;
 import com.wurmcraft.serveressentials.forge.common.utils.PlayerUtils;
 import com.wurmcraft.serveressentials.forge.common.utils.TextComponentUtils;
 import com.wurmcraft.serveressentials.forge.modules.general.GeneralConfig;
+import com.wurmcraft.serveressentials.forge.modules.rank.RankUtils;
 import java.util.NoSuchElementException;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.entity.player.EntityPlayer;
@@ -28,43 +31,53 @@ public class DelHomeCommand {
 
   @Command(inputArguments = {CommandArguments.STRING}, inputNames = {"Home"})
   public void delHome(ICommandSender sender, String home) {
-    if (sender != null && sender.getCommandSenderEntity() != null && sender
-        .getCommandSenderEntity() instanceof EntityPlayer) {
-      EntityPlayer player = (EntityPlayer) sender.getCommandSenderEntity();
-      try {
-        StoredPlayer playerData = (StoredPlayer) SERegistry
-            .getStoredData(DataKey.PLAYER, player.getGameProfile().getId().toString());
-        if (home.equalsIgnoreCase("all")) {
-          playerData.server.homes = new Home[0];
-          SERegistry.register(DataKey.PLAYER, playerData);
-          sender.sendMessage(new TextComponentString(PlayerUtils.getUserLanguage(sender).GENERAL_DELHOME_ALL));
-        } else {
-          for (Home h : playerData.server.homes) {
-            if (h.name.equals(home)) {
-              SERegistry
-                  .register(DataKey.PLAYER, PlayerUtils.deleteHome(playerData, home));
-              sender.sendMessage(TextComponentUtils.addPosition(new TextComponentString(
-                  PlayerUtils.getUserLanguage(sender).GENERAL_DELHOME_REMOVED
-                      .replaceAll("%HOME%", h.name)), h));
-              return;
+    if (SERegistry.isModuleLoaded("Rank") && RankUtils
+        .hasPermission(RankUtils.getRank(sender), "general.delhome") || !SERegistry
+        .isModuleLoaded("Rank")) {
+      if (sender != null && sender.getCommandSenderEntity() != null && sender
+          .getCommandSenderEntity() instanceof EntityPlayer) {
+        EntityPlayer player = (EntityPlayer) sender.getCommandSenderEntity();
+        try {
+          StoredPlayer playerData = (StoredPlayer) SERegistry
+              .getStoredData(DataKey.PLAYER, player.getGameProfile().getId().toString());
+          if (home.equalsIgnoreCase("all")) {
+            playerData.server.homes = new Home[0];
+            SERegistry.register(DataKey.PLAYER, playerData);
+            sender.sendMessage(new TextComponentString(
+                COMMAND_COLOR + PlayerUtils.getUserLanguage(sender).GENERAL_DELHOME_ALL));
+          } else {
+            for (Home h : playerData.server.homes) {
+              if (h.name.equals(home)) {
+                SERegistry
+                    .register(DataKey.PLAYER, PlayerUtils.deleteHome(playerData, home));
+                sender.sendMessage(
+                    TextComponentUtils.addPosition(new TextComponentString(COMMAND_COLOR +
+                        PlayerUtils.getUserLanguage(sender).GENERAL_DELHOME_REMOVED
+                            .replaceAll("%HOME%", COMMAND_INFO_COLOR + h.name)), h));
+                return;
+              }
             }
-          }
-          for (Home h : playerData.server.homes) {
-            if (h.name.equalsIgnoreCase(home)) {
-              SERegistry
-                  .register(DataKey.PLAYER, PlayerUtils.deleteHome(playerData, home));
-              sender.sendMessage(TextComponentUtils.addPosition(new TextComponentString(
-                  PlayerUtils.getUserLanguage(sender).GENERAL_DELHOME_REMOVED
-                      .replaceAll("%HOME%", h.name)), h));
-              return;
+            for (Home h : playerData.server.homes) {
+              if (h.name.equalsIgnoreCase(home)) {
+                SERegistry
+                    .register(DataKey.PLAYER, PlayerUtils.deleteHome(playerData, home));
+                sender.sendMessage(
+                    TextComponentUtils.addPosition(new TextComponentString(COMMAND_COLOR +
+                        PlayerUtils.getUserLanguage(sender).GENERAL_DELHOME_REMOVED
+                            .replaceAll("%HOME%", COMMAND_INFO_COLOR + h.name)), h));
+                return;
+              }
             }
+            sender.sendMessage(new TextComponentString(
+                ERROR_COLOR + PlayerUtils.getUserLanguage(sender).GENERAL_HOME_NOT_FOUND
+                    .replaceAll("%HOME%", COMMAND_INFO_COLOR + home)));
           }
-          sender.sendMessage(new TextComponentString(
-              ERROR_COLOR + PlayerUtils.getUserLanguage(sender).GENERAL_HOME_NOT_FOUND
-                  .replaceAll("%HOME%", home)));
+        } catch (NoSuchElementException ignored) {
         }
-      } catch (NoSuchElementException ignored) {
       }
+    } else {
+      sender.sendMessage(new TextComponentString(
+          ERROR_COLOR + PlayerUtils.getUserLanguage(sender).ERROR_NO_PERMS));
     }
   }
